requirements.txt
langchain>=0.3.27
langchain-community>=0.3.27
langchain-core>=0.3.72
langchain-huggingface>=0.3.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pypdf>=5.8.0,<7.0.0
openai>=1.40.0,<2.0.0
chromadb==1.0.15
sentence-transformers>=3.0.0,<4.0.0
transformers>=4.39.0
tiktoken>=0.8.0,<1.0.0
numpy>=2.1.0
pandas>=2.2.0,<3.0.0
faiss-cpu>=1.7.4,<2.0.0
pydantic>=2.7.4,<3.0.0
httpx>=0.27.0,<1.0.0
tqdm>=4.67.0,<5.0.0
beautifulsoup4>=4.12.3,<5.0.0
requests>=2.31.0,<3.0.0
ipywidgets>=8.0.0,<9.0.0
urllib3>=2.2.1,<3.0.0
streamlit
watchdog
langchain-openai>=0.3.28
langchain-chroma>=0.3.0